M54 Rebasing: Move Creation of GeolocationDelegate to the embedders
adapt to
https://codereview.chromium.org/2174613002/
https://codereview.chromium.org/2184013004
